refactor bessel solver
